backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
argon2-cffi
gunicorn